Refactor TFPolicy and Policy
