+++ ml-agents/mlagents/trainers/tests/test_nn_policy.py
@@ -8,7 +8,7 @@ from mlagents.model_serialization import SerializationSettings
 from mlagents.tf_utils import tf
 
 
-from mlagents.trainers.policy.nn_policy import NNPolicy
+from mlagents.trainers.policy.tf_policy import TFPolicy
 from mlagents.trainers.models import EncoderType, ModelUtils, Tensor3DShape
 from mlagents.trainers.exception import UnityTrainerException
 from mlagents.trainers.tests import mock_brain as mb
@@ -32,7 +32,7 @@ def create_policy_mock(
     model_path: str = "",
     load: bool = False,
     seed: int = 0,
-) -> NNPolicy:
+) -> TFPolicy:
     mock_spec = mb.setup_test_behavior_specs(
         use_discrete,
         use_visual,
@@ -47,7 +47,9 @@ def create_policy_mock(
     trainer_settings.network_settings.memory = (
         NetworkSettings.MemorySettings() if use_rnn else None
     )
-    policy = NNPolicy(seed, mock_spec, trainer_settings, False, model_path, load)
+    policy = TFPolicy(
+        seed, mock_spec, trainer_settings, model_path=model_path, load=load
+    )
     return policy
 
 
@@ -101,7 +103,7 @@ class ModelVersionTest(unittest.TestCase):
             assert len(cm.output) == 1
 
 
-def _compare_two_policies(policy1: NNPolicy, policy2: NNPolicy) -> None:
+def _compare_two_policies(policy1: TFPolicy, policy2: TFPolicy) -> None:
     """
     Make sure two policies have the same output for the same input.
     """
@@ -149,11 +151,10 @@ def test_normalization():
     # Change half of the obs to 0
     for i in range(3):
         trajectory.steps[i].obs[0] = np.zeros(1, dtype=np.float32)
-    policy = NNPolicy(
+    policy = TFPolicy(
         0,
         behavior_spec,
         TrainerSettings(network_settings=NetworkSettings(normalize=True)),
-        False,
         "testdir",
         False,
     )
